Move SearchResultGeneration into the package which uses it
